+++ NEWS.md
@@ -121,6 +121,7 @@ argument systematically anymore
 * Add the remove.missing argument in report.quali to be able to compute the percentage on non missing values only if needed
 * Add transpose function
 * Add hierarchical descriptive statistics for reporting Adverse Event, Medical History and concomitant treatment
+* Add report.sessionInfo to export to Word or R markdown do, the sessionInfo of R
 
 ## Issues
 * Correction of a bug in report.quanti when x1 and x2 arguments were NULL